+++ lib/src/foundation/nsnumber.dart
@@ -12,8 +12,8 @@ class NSNumber extends NSValue {
     String typeName = value.runtimeType.toString();
     if (_NSNumberCreationForBoxType.containsKey(typeName)) {
       String selName = 'numberWith${_NSNumberCreationForBoxType[typeName]}:';
-      NSObject result = type(of: NSNumber)
-        .perform(Selector(selName), args: [value]);
+      NSObject result =
+          type(of: NSNumber).perform(Selector(selName), args: [value]);
       return result.pointer;
     } else {
       throw 'Unknown type for initializing NSNumber.';